requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.0.5
django-ckeditor==6.0.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-phonenumber-field==5.1.0
djongo==1.3.4
gunicorn==20.1.0
idna==2.10
numpy==1.20.2
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
phonenumbers==8.12.22
Pillow==8.2.0
psycopg2==2.8.6
pymongo==3.11.3
pytz==2021.1
requests==2.25.1
sqlparse==0.2.4
urllib3==1.26.4
whitenoise==5.2.0
wincertstore==0.2
